[ADD] Api & Frontend HTTP Requests
Frontend:
- CRUD from projects api;
- CSS styling for the sake of browser readability ;

Note: When you press "Settings" to use the "UPDATE" request, there's a minor bug that opens a form on all projects athough it only updates the selected project. It will be fixed on next commit along with components refactor & routing.

Api:
- Added CRUD operations for projects model;

Note: Video & Photo fields are still as a string and will be fixed on next commit along with code refactor.